requirements.txt
opencv-python
dlib-bin
imutils
scipy
playsound
